Using size_t for index
1. Using size_t for index
2. Added TreeNode

+++ ukive/views/list/flow_list_layouter.h
@@ -19,8 +19,8 @@ namespace ukive {
 
         void onMeasureAtPosition(bool cur, int width, int height) override;
         int onLayoutAtPosition(bool cur) override;
-        int onScrollToPosition(int pos, int offset, bool cur) override;
-        int onSmoothScrollToPosition(int pos, int offset) override;
+        int onScrollToPosition(size_t pos, int offset, bool cur) override;
+        int onSmoothScrollToPosition(size_t pos, int offset) override;
 
         int onFillTopChildren(int dy) override;
         int onFillBottomChildren(int dy) override;
@@ -34,23 +34,23 @@ namespace ukive {
         ListItem* findItemFromView(View* v) override;
 
         bool canScroll(Direction dir) const override;
-        void getCurPosition(int* pos, int* offset) const override;
+        void getCurPosition(size_t* pos, int* offset) const override;
 
     private:
         struct Record {
-            int cur_row = 0;
+            size_t cur_row = 0;
             int cur_offset = 0;
             bool is_null = true;
         };
 
-        int calPreferredCurPos() const;
+        size_t calPreferredCurPos() const;
 
         bool canScrollToTop() const;
         bool canScrollToBottom() const;
         bool canScrollToLeft() const;
         bool canScrollToRight() const;
 
-        int col_count_;
+        size_t col_count_;
         ColumnCollection columns_;
         std::vector<Record> cur_records_;
     };
